+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.2] - 2020-03-24
+
+Improve handling of parse-errors.
+
+Fix default format for files (was and is documented to be uncolored, but was colored).
+
 ## [0.15.1] - 2020-03-04
 
 Make the textfilter functionality an optional default feature;